// src/api/enum.ts
enum ROLE {
  ADMIN = 'ADMIN',
  WAREHOUSE = 'WAREHOUSE',
  DELIVERY = 'DELIVERY',
  MAKER = 'MAKER'
}
enum STATE {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE'
}
enum STATE_MAINTENANCE {
  DOING = 'DOING',
  COMPLETED = 'COMPLETED',
  DELETED = 'DELETED'
}
enum STATE_MAINTAIN {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  MAINTAIN = 'MAINTAIN'
}
enum TYPE_MACHINE {
  FIXED = 'FIXED',
  MOVING = 'MOVING'
}
enum TYPE_WAREHOUSE {
  MAKER = 'MAKER',
  HONDA = 'HONDA',
  OTHER = 'OTHER'
}
enum STATUS_TRANSFER {
  FIND = 'FIND',
  SCANNING = 'SCANNING',
  TRANSFER = 'TRANSFER',
  SUCCESS = 'SUCCESS'
}
enum STATUS_TRANSFERDETAIL {
  CANCEL = 'CANCEL',
  SOLVE = 'SOLVE',
  TRANSPORT = 'TRANSPORT',
  SUCCESS = 'SUCCESS'
}
enum IMPORT_SCAN {
  SCAN = 'SCAN',
  ENTER = 'ENTER'
}

enum MAINTAIN_TIME {
  DAY_15 = 86400000 * 15,
  DAY_30 = 86400000 * 30,
  DAY_60 = 86400000 * 60,
  DAY_90 = 86400000 * 90
}

export {
  ROLE,
  STATE,
  STATE_MAINTAIN,
  STATE_MAINTENANCE,
  TYPE_MACHINE,
  TYPE_WAREHOUSE,
  STATUS_TRANSFER,
  STATUS_TRANSFERDETAIL,
  IMPORT_SCAN,
  MAINTAIN_TIME
}